Move layout into a component
