+++ src/components/__tests__/Layout.test.js
@@ -27,40 +27,40 @@ const props = {
 
 describe('index layout page', () => {
   it('renders without error', () => {
-    const IndexLayout = require('../index').default
+    const IndexLayout = require('../Layout').default
     shallow(<IndexLayout {...props} />)
   })
 
   it('sets the canonical URL', () => {
-    const IndexLayout = require('../index').default
+    const IndexLayout = require('../Layout').default
     const wrapper = shallow(<IndexLayout {...props} />)
     const elem = wrapper.find('link[rel="canonical"]')
     expect(elem.prop('href')).toBe('https://tab.gladly.io/')
   })
 
   it('sets the default page title using Helmet', () => {
-    const IndexLayout = require('../index').default
+    const IndexLayout = require('../Layout').default
     const wrapper = shallow(<IndexLayout {...props} />)
     const elem = wrapper.find(Helmet)
     expect(elem.prop('defaultTitle')).toBe('My Example Site')
   })
 
   it('sets the open graph title to the call-to-action text', () => {
-    const IndexLayout = require('../index').default
+    const IndexLayout = require('../Layout').default
     const wrapper = shallow(<IndexLayout {...props} />)
     const elem = wrapper.find('meta[property="og:title"]')
     expect(elem.prop('content')).toBe('Check this out!')
   })
 
   it('sets the open graph description', () => {
-    const IndexLayout = require('../index').default
+    const IndexLayout = require('../Layout').default
     const wrapper = shallow(<IndexLayout {...props} />)
     const elem = wrapper.find('meta[property="og:description"]')
     expect(elem.prop('content')).toBe('A shorter description')
   })
 
   it('sets the open graph image', () => {
-    const IndexLayout = require('../index').default
+    const IndexLayout = require('../Layout').default
     const wrapper = shallow(<IndexLayout {...props} />)
     const elem = wrapper.find('meta[property="og:image"]')
     expect(elem.prop('content')).toBe(
